requirements.txt
streamlit
fastapi
uvicorn
python-multipart
torch
torchvision
Pillow